backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.16.0
numpy>=1.24.0
Pillow>=10.0.0
keras>=2.15.0
h5py>=3.9.0
python-dotenv==1.0.0
itsdangerous==2.1.2
bcrypt==4.1.2
torch==2.10.0
torchvision==0.25.0
